requirements.txt
requests>=2.31.0
pynput>=1.7.6
pyserial>=3.5

